requirements.txt
asyncpg
python-dotenv
aiogram